add replacement for prism2 header - see extradecap, extraencap. This allows click to run at userlevel and still do rate control, and get feedback that used to be only available in the kernel module.

+++ elements/wifi/probetxrate.hh
@@ -291,7 +291,7 @@ class ProbeTXRate : public Element { public:
 	  continue;
 	}
 
-	if (_total_success[x] + _total_fail[x] < min_sample) {
+	if (_total_success[x] + _total_fail[x] < (int) min_sample) {
 	  possible_rates.push_back(_rates[x]);
 	}
       }
